+++ src/main/java/epermit/data/repositories/AuthorityKeyRepository.java
@@ -6,6 +6,6 @@ import epermit.data.entities.AuthorityKey;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface AuthorityKeyRepository extends JpaRepository<AuthorityKey, Long> {
+public interface AuthorityKeyRepository extends JpaRepository<AuthorityKey, Integer> {
    
 }
