+++ src/main/java/epermit/data/repositories/AuthorityQuotaRepository.java
@@ -6,6 +6,6 @@ import epermit.data.entities.AuthorityQuota;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface AuthorityQuotaRepository extends JpaRepository<AuthorityQuota, Long> {
+public interface AuthorityQuotaRepository extends JpaRepository<AuthorityQuota, Integer> {
     
 }
