+++ src/main/java/epermit/data/repositories/KeyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 @Repository
-public interface KeyRepository extends JpaRepository<Key, Long> {
+public interface KeyRepository extends JpaRepository<Key, Integer> {
      //Key findFirstByOrderByIdDesc();
      @Query(value = "select k from Key k where k.enabled = true")
      Key getEnabled();
